+++ microgen/shape/strut_lattice/truncated_octahedron.py
@@ -29,7 +29,7 @@ class TruncatedOctahedron(AbstractLattice):
     def _compute_vertices(self) -> npt.NDArray[np.float64]:
         return self.center + self.cell_size * self._base_vertices
     
-    def _generate_strut_vertex_pairs(self) -> npt.NDArray[int]:
+    def _generate_strut_vertex_pairs(self) -> npt.NDArray[np.int64]:
         """Generate index pairs representing the struts in the truncated octahedron."""
         tree = KDTree(self._base_vertices)
         tolerance = 1e-5
